Fix NullPointerException for onSinglePointerTouchEvent
NullPointerException occurs when trying to get key.valueText
for onSinglePointerTouchEvent of a null key,
which occurs if the keyboard rectangle is not completely filled with keys
and the user drags a finger over a point that is not inside a key.
---
After implementing anort on (key == null),
note that we will automatically satisfy the
'cancel key press if move outside keyboard' behaviour
by putting in a candidates bar at the top of the keyboard.