Implement abort on move outside keyboard
by adding a 1 px gutter at the top of the keyboard.
The fix in c40fbe2 already makes typing abort
when the user moves onto a pixel not occupied by a key.
But that only works for pixels inside the InputContainer.
This commit ensures that there is a 1 px row
which is key-free, at the top of our InputContainer.