Create new class KeyPreviewPlane extending View
The existing implementation of key preview popups is laggy and crappy,
as discussed in the merge commit 8b17b22.
This commit begins a new approach:
instead of having KeyPreview being a View with dimensions equal to the popup
(which needs to be resized and repositioned whenever activeKey changes),
we have a KeyPreviewPlane being a View that covers the whole keyboard.