Eliminate slow loop in computeSortingRank
Old version was extremely slow because it did a String.equals comparison
inside a for loop over the list of phrase comparisons.
This loop is now eliminated, by storing a list of first characters
and doing a lookup (via indexOf),
instead of extracting the first character inside the loop
(which is extremely dumb).