Use popup recess height for preview vertical positioning
This is a temporary hack.
Basically, the issue is that key preview plane is anchored
to the top of the input container (however tall the popup recess is).
This commit simply uses the popup recess height
for the vertical positioning of the key previews.
This should be edited to account for the actual height discrepancy
so that the preview margin doesn't look too big
for the bottom row of keys.